requirements.txt
aiohttp==3.4.4
async-timeout==3.0.1
attrs==18.2.0
chardet==3.0.4
git+git://github.com/Rapptz/discord.py@3f06f247c039a23948e7bb0014ea31db533b4ba2
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
python-dotenv==0.10.1
websockets==6.0
yarl==1.3.0